use nsenter to manage ebs mounts from container
