+++ engine/src/juliabox/plugins/vol_ebs/ebs.py
@@ -1,10 +1,14 @@
 import os
 import threading
 import time
+#import sh
+from string import ascii_lowercase
 
+from juliabox.cloud.awsebsvol import EBSVol
 from juliabox.cloud.aws import CloudHost
 from juliabox.db import JBoxSessionProps
 from juliabox.jbox_util import unique_sessname, JBoxCfg
+# from juliabox.jbox_util import create_host_mnt_command
 from juliabox.vol import JBoxVol
 from disk_state_tbl import JBoxDiskState
 from juliabox.jbox_container import JBoxContainer
@@ -21,18 +25,26 @@ class JBoxEBSVol(JBoxVol):
     DISK_RESERVE_TIME = {}
     DISK_TEMPLATE_SNAPSHOT = None
     LOCK = None
+    # SH_READ_FSTAB = None
 
     @staticmethod
     def configure():
+        # JBoxEBSVol.SH_READ_FSTAB = create_host_mnt_command("cat /etc/fstab")
         num_disks_max = JBoxCfg.get('numdisksmax')
         JBoxEBSVol.FS_LOC = os.path.expanduser(JBoxCfg.get('cloud_host.ebs_mnt_location'))
         JBoxEBSVol.DISK_LIMIT = 1
         JBoxEBSVol.MAX_DISKS = num_disks_max
         JBoxEBSVol.DISK_TEMPLATE_SNAPSHOT = JBoxCfg.get('cloud_host.ebs_template')
 
-        JBoxEBSVol.DEVICES = JBoxEBSVol._get_configured_devices(JBoxEBSVol.FS_LOC)
-        if len(JBoxEBSVol.DEVICES) < num_disks_max:
-            raise Exception("Not enough EBS mount points configured")
+        JBoxEBSVol.DEVICES = JBoxEBSVol._guess_configured_devices('xvd', num_disks_max)
+        JBoxEBSVol.log_debug("Assuming %d EBS volumes configured in range xvdba..xvdcz", len(JBoxEBSVol.DEVICES))
+        # JBoxEBSVol.DEVICES = JBoxEBSVol._get_configured_devices(JBoxEBSVol.FS_LOC)
+        # if len(JBoxEBSVol.DEVICES) < num_disks_max:
+        #     JBoxEBSVol.log_error("Not enough EBS mount points configured. Found %d, configured %d",
+        #                          len(JBoxEBSVol.DEVICES), num_disks_max)
+        #     raise Exception("Not enough EBS mount points configured")
+        # else:
+        #     JBoxEBSVol.log_debug("Found %d EBS volumes configured", len(JBoxEBSVol.DEVICES))
 
         JBoxEBSVol.LOCK = threading.Lock()
         JBoxEBSVol.refresh_disk_use_status()
@@ -41,22 +53,31 @@ class JBoxEBSVol(JBoxVol):
     def get_disk_allocated_size(cls):
         return JBoxEBSVol.DISK_LIMIT * 1000000000
 
-    @staticmethod
-    def _id_from_device(dev_path):
-        return dev_path.split('/')[-1]
+    # @staticmethod
+    # def _id_from_device(dev_path):
+    #     return dev_path.split('/')[-1]
+    #
+    # @staticmethod
+    # def _get_configured_devices(fs_loc):
+    #     devices = []
+    #     for line in JBoxEBSVol.SH_READ_FSTAB():
+    #         line = line.strip()
+    #         if (len(line) == 0) or line.startswith('#'):
+    #             continue
+    #         comps = line.split()
+    #         if (len(comps) == 6) and comps[1].startswith(fs_loc):
+    #             device = comps[0]
+    #             devices.append(JBoxEBSVol._id_from_device(device))
+    #     return devices
 
     @staticmethod
-    def _get_configured_devices(fs_loc):
+    def _guess_configured_devices(devidpfx, num_disks):
         devices = []
-        with open('/etc/fstab', 'r') as fstab:
-            for line in fstab:
-                line = line.strip()
-                if (len(line) == 0) or line.startswith('#'):
-                    continue
-                comps = line.split()
-                if (len(comps) == 6) and comps[1].startswith(fs_loc):
-                    device = comps[0]
-                    devices.append(JBoxEBSVol._id_from_device(device))
+        for pfx1 in 'bc':
+            for pfx2 in ascii_lowercase:
+                devices.append(devidpfx+pfx1+pfx2)
+                if len(devices) == num_disks:
+                    return devices
         return devices
 
     @staticmethod
@@ -166,17 +187,15 @@ class JBoxEBSVol(JBoxVol):
 
             JBoxEBSVol.log_debug("will use snapshot id %s for %s", snap_id, user_email)
 
-            _dev_path, mnt_path, vol_id = CloudHost.create_new_volume(snap_id, disk_id,
-                                                                      JBoxEBSVol.FS_LOC,
-                                                                      tag=user_email,
-                                                                      disk_sz_gb=JBoxEBSVol.DISK_LIMIT)
+            _dev_path, mnt_path, vol_id = EBSVol.create_new_volume(snap_id, disk_id, JBoxEBSVol.FS_LOC,
+                                                                   tag=user_email, disk_sz_gb=JBoxEBSVol.DISK_LIMIT)
             existing_disk = JBoxDiskState(cluster_id=CloudHost.INSTALL_ID, region_id=CloudHost.REGION,
                                           user_id=user_email,
                                           volume_id=vol_id,
                                           attach_time=None,
                                           create=True)
         else:
-            _dev_path, mnt_path = CloudHost.attach_volume(existing_disk.get_volume_id(), disk_id, JBoxEBSVol.FS_LOC)
+            _dev_path, mnt_path = EBSVol.attach_volume(existing_disk.get_volume_id(), disk_id, JBoxEBSVol.FS_LOC)
             existing_disk.set_attach_time()
             snap_id = None
 
@@ -220,24 +239,23 @@ class JBoxEBSVol(JBoxVol):
         sess_props = JBoxSessionProps(self.sessname)
         desc = sess_props.get_user_id() + " JuliaBox Backup"
         disk_id = self.disk_path.split('/')[-1]
-        snap_id = CloudHost.snapshot_volume(dev_id=disk_id, tag=self.sessname, description=desc,
-                                            wait_till_complete=False)
-        #old_snap_id = sess_props.get_snapshot_id()
-        #sess_props.set_snapshot_id(snap_id)
-        #sess_props.save()
-        #if old_snap_id is not None:
+        snap_id = EBSVol.snapshot_volume(dev_id=disk_id, tag=self.sessname, description=desc, wait_till_complete=False)
+        # old_snap_id = sess_props.get_snapshot_id()
+        # sess_props.set_snapshot_id(snap_id)
+        # sess_props.save()
+        # if old_snap_id is not None:
         #    CloudHost.delete_snapshot(old_snap_id)
         return snap_id
 
     def release(self, backup=False):
         disk_id = self.disk_path.split('/')[-1]
-        CloudHost.unmount_device(disk_id, JBoxEBSVol.FS_LOC)
+        EBSVol.unmount_device(disk_id, JBoxEBSVol.FS_LOC)
         if backup:
             snap_id = self._backup()
         else:
             snap_id = None
-        vol_id = CloudHost.get_volume_id_from_device(disk_id)
-        CloudHost.detach_volume(vol_id, delete=False)
+        vol_id = EBSVol.get_volume_id_from_device(disk_id)
+        EBSVol.detach_volume(vol_id, delete=False)
 
         sess_props = JBoxSessionProps(self.sessname)
         existing_disk = JBoxDiskState(cluster_id=CloudHost.INSTALL_ID, region_id=CloudHost.REGION,
