+++ host/supervisord.conf
@@ -26,7 +26,7 @@ stderr_logfile_backups = 2
 stderr_logfile_maxbytes = 1MB
 
 [program:enginedaemon]
-command = docker run -a stdout -a stderr --rm --name=enginedaemon_jboxsvc --privileged --net=host -v /var/run/docker.sock:/var/run/docker.sock -v /jboxengine/data:/jboxengine/data -v %(here)s/../engine/logs:/jboxengine/logs juliabox/enginedaemon
+command = docker run -a stdout -a stderr --rm --name=enginedaemon_jboxsvc --privileged --net=host -v /proc:/hostproc -v /var/run/docker.sock:/var/run/docker.sock -v /jboxengine/data:/jboxengine/data -v %(here)s/../engine/logs:/jboxengine/logs juliabox/enginedaemon
 directory =  %(here)s/../engine
 process_name = enginedaemon
 stdout_logfile = %(here)s/../engine/logs/enginedaemon.log
